Turn quality settings way up
These are settings that more or less guarantee anything you find
with reasonable probability you find in a test. This means we're
testing for things you can find over multiple runs rather than for
things you find over a single one.

Because this is intended to be run with example saving this is really
the right thing to optimise for.