Initial bulk byte simplify.
Try to do a shared simplification by mapping all bytes of a shared
value to some smaller value.

This seems to improve some cases where we'd have to do a lot of
little simplifications bytewise by some margin. For some reason it
helps a lot with the floating point tests.